tudo puta!!!

+++ parallelMpi/physics.h
@@ -1,8 +1,8 @@
 // physics.h
 #include "grid.h"
 
-void divide_par(int n_pr, int part_no, int rem, int *par_block);
-int get_par_number(int part_no, int *par_block, int n_rank, int n_pr);
+void divide_par(int n_pr, long long part_no, int rem, int *par_block);
+long long get_par_number(long long part_no, int *par_block, int n_rank, int n_pr);
 int task_owner(int part_no, int n_pr, int rem, int par);
 double init_particles(long seed, long ncside, long long n_part, Particle *par, Grid **grid, int *par_block, int n_pr);
 int fill_par_buffer(double *par_buffer, Particle *par, int aux_i, int pr_part_no);
